+++ src/Systems/Peer.h
@@ -327,8 +327,8 @@ public:
 	virtual void setFirmwareVersion(int32_t value) { _firmwareVersion = value; saveVariable(1001, value); }
 	virtual std::string getFirmwareVersionString() { return _firmwareVersionString; }
 	virtual void setFirmwareVersionString(std::string value) { _firmwareVersionString = value; saveVariable(1003, value); }
-	virtual uint32_t getDeviceType() { return _deviceType; }
-	virtual void setDeviceType(uint32_t value) { _deviceType = value; saveVariable(1002, (int32_t)_deviceType); initializeTypeString(); }
+	virtual uint64_t getDeviceType() { return _deviceType; }
+	virtual void setDeviceType(uint64_t value) { _deviceType = value; saveVariable(1002, (int64_t)_deviceType); initializeTypeString(); }
     virtual std::string getName() { return getName(-1); }
 	virtual std::string getName(int32_t channel);
 	virtual void setName(std::string value) { setName(-1, value); }
@@ -498,7 +498,7 @@ protected:
 	//In table variables:
 	int32_t _firmwareVersion = 0;
 	std::string _firmwareVersionString;
-	uint32_t _deviceType = 0;
+	uint64_t _deviceType = 0;
 	std::mutex _peersMutex;
 	std::unordered_map<int32_t, std::vector<std::shared_ptr<BasicPeer>>> _peers;
 	std::mutex _namesMutex;
